src/app/pyproject.toml
[tool.poetry]
name = "rankkeeper"
version = "0.1.0"
description = ""
authors = ["kousakirai <kouki.muro2850@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.9.19"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
motor = "^3.3.2"
sqlalchemy = "^2.0.29"
aiomysql = "^0.2.0"
jishaku = "^2.5.2"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
